requirements.txt
termcolor
requests
aiohttp
pynacl
colorama